creación de BotEquipo5 y abstracción de Bot
