+++ Model/Bot.java
@@ -6,9 +6,9 @@ import java.util.Arrays;
 import java.util.Comparator;
 import java.util.Stack;
 
-public class Bot extends Jugador {
-    private ControladorEstructura controladorEstructura;
-    private Casilla casilla;
+public abstract class Bot extends Jugador {
+    protected ControladorEstructura controladorEstructura;
+    protected Casilla casilla;
     //   private ArrayList<Nodo> arregloNodos;
     // Algoritmo algoritmo;
 
@@ -23,13 +23,14 @@ public class Bot extends Jugador {
         this.casilla = null;
 
         this.controladorEstructura = new ControladorEstructura(controlador);
+        this.analizarEstrategia();
         this.controladorEstructura.getEstructuraColumnas().sort(new Comparator<Estructura>() {
             @Override
             public int compare(Estructura o1, Estructura o2) {
                 return o2.getCola().size() - o1.getCola().size();
             }
         });
-        this.formarQwirkle();
+        this.analizarEstrategia();
     /*    if (casilla == null && this.arregloNodos.isEmpty()) {
             Nodo nodo = new Nodo(this.controladorEstructura, this.getFichas2());
             Algoritmo algoritmo = new Algoritmo(nodo);
@@ -46,105 +47,11 @@ public class Bot extends Jugador {
 
     }
 
+    public abstract void analizarEstrategia() throws Exception;
+
     public Casilla getCasilla() {
         return casilla;
     }
 
 
-    public void formarQwirkle() throws Exception {
-        for (int i = 5; i > 0; i--) {
-            if (this.casilla == null) {
-                if (this.isJugadaEncontradaFilas(i)) {
-                    // JOptionPane.showMessageDialog(null, "HE ENCONTRADO FICHA Y CASILLA");
-                } else {
-                    if (this.isJugadaEncontradaColumnas(i)) {
-                        // JOptionPane.showMessageDialog(null, "HE ENCONTRADO FICHA Y CASILLA");
-                    }
-                }
-            } else {
-                break;
-            }
-        }
-    }
-
-    public boolean isJugadaEncontradaFilas(int i) throws Exception {
-        for (Estructura fila : this.controladorEstructura.getEstructuraFilas()) {
-            if (fila.getCola().size() == i) {
-                for (Ficha ficha : this.getArregloFichas()) {
-                    if (Regla.cumpleRestriccion(fila, ficha)) {
-                        if (fila.getCola().getLast().getCasillaDerecha() != null) {
-                            try {
-                                this.controladorEstructura.agregar(fila.getCola().getLast().getCasillaDerecha(), ficha);
-                                this.setFichaSeleccionada(ficha);
-                                fila.getCola().getLast().setFicha(null);
-                                this.casilla = fila.getCola().getLast();
-                                return true;
-                            } catch (Exception e) {
-                                //   e.printStackTrace();
-                                System.out.println("revienta el juegoooo");
-                            }
-                        } else {
-                            if (fila.getCola().getFirst().getCasillaIzquierda() != null) {
-                                try {
-                                    this.controladorEstructura.agregar(fila.getCola().getFirst().getCasillaIzquierda(), ficha);
-                                    this.setFichaSeleccionada(ficha);
-                                    fila.getCola().getFirst().setFicha(null);
-                                    this.casilla = fila.getCola().getFirst();
-                                    this.casilla.setFicha(null);
-                                    return true;
-                                } catch (Exception e) {
-                                    //    e.printStackTrace();
-                                    System.out.println("revienta el juegoooo");
-                                }
-                            }
-
-                        }
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    public boolean isJugadaEncontradaColumnas(int i) throws Exception {
-        for (Estructura columna : this.controladorEstructura.getEstructuraColumnas()) {
-            if (columna.getCola().size() == i) {
-                for (Ficha ficha : this.getArregloFichas()) {
-                    if (Regla.cumpleRestriccion(columna, ficha)) {
-                        if (columna.getCola().getLast().getCasillaInferior() != null) {
-                            try {
-                                this.controladorEstructura.agregar(columna.getCola().getLast().getCasillaInferior(), ficha);
-                                this.setFichaSeleccionada(ficha);
-
-                                columna.getCola().getLast().setFicha(null);
-                                this.casilla = columna.getCola().getLast();
-                                this.casilla.setFicha(null);
-                                return true;
-                            } catch (Exception e) {
-                                // e.printStackTrace();
-                                System.out.println("revienta el juegoooo");
-                            }
-                        } else {
-                            if (columna.getCola().getFirst().getCasillaSuperior() != null) {
-                                try {
-                                    this.controladorEstructura.agregar(columna.getCola().getFirst().getCasillaSuperior(), ficha);
-                                    this.setFichaSeleccionada(ficha);
-                                    columna.getCola().getFirst().setFicha(null);
-                                    this.casilla = columna.getCola().getFirst();
-                                    this.casilla.setFicha(null);
-
-                                    return true;
-                                } catch (Exception e) {
-                                    //   e.printStackTrace();
-                                    System.out.println("revienta el juegoooo");
-                                }
-                            }
-                        }
-                    }
-
-                }
-            }
-        }
-        return false;
-    }
 }
